LangChain/pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.1",
    "bs4>=0.0.2",
    "chromadb>=1.5.5",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langserve>=0.3.3",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "sse-starlette>=3.3.3",
    "uvicorn>=0.42.0",
    "wikipedia>=1.4.0",
]
